refactor(forwarder): move outdated retry file cleanup into PersistedQueue::from_root_path
Cleanup now lives inside saluki-io alongside the filename format it depends
on. No cross-crate exports needed. RetryQueue::with_disk_persistence and
PersistedQueue::from_root_path each gain a max_age_days: u32 parameter;
io.rs passes outdated_file_in_days() at the single call site.
