+++ lib/saluki-io/src/net/util/retry/mod.rs
@@ -11,9 +11,7 @@ mod policy;
 pub use self::policy::{NoopRetryPolicy, RollingExponentialBackoffRetryPolicy};
 
 mod queue;
-pub use self::queue::{
-    decode_timestamped_filename, DiskUsageRetrieverImpl, EventContainer, PushResult, RetryQueue, Retryable,
-};
+pub use self::queue::{DiskUsageRetrieverImpl, EventContainer, PushResult, RetryQueue, Retryable};
 
 /// A batteries-included retry policy suitable for HTTP-based clients.
 pub type DefaultHttpRetryPolicy<B = ()> =
